Fix/update the `btcinfo` bash alias
as per current Bitcoin master.

`getwalletinfo` appears to no longer be part of bitcoin-cli, but we can obtain the balance as well as the version and number of connections with `-getinfo` instead.

This commit also improves the terminal output for more consistent coloring of the double quotes.

Further suggestions welcome.