Update 03_2_Knowing_Your_Bitcoin_Setup.md
I was having some problems getting the curl command to work and if you aren't familiar with how RPC works it can be quite confusing (what is an RPC user and password - is it the same as my username and password?). 

For some reason I was able to get it working by specifying the user and the password in the HTTP request - which I think is clearer (also how fullyNoded does it). 

Learning how to use curl effectively can be quite useful so becoming familiar with it can be valuable to a budding developer.

I also added a pipe to jq and told it to pretty print everything otherwise its not as readable.

Explained that in the HTTP request, bitcoin core knows to reply with testnet data (or mainnet) because the HTTP request is asking for a resource that can only be accessed through that port. It's implicit, not explicit what curl is requesting (or at least IMO this is how I see it). On my machine - 18883 is being used instead of 8883, luckily I know how to see which port a process is using.