Initial Setup, Step 4, increased clarity on version
It is not clear until chapter 3 that the remainder of the tutorial will assume we chose a pruned testnet. When I first did the setup I used a pruned mainnet, and thus by chapter 3 I had to rebuild the node so that I could better follow along. In retrospect it might seem apparent from the warning about using real funds in earlier paragraphs, but to reduce any confusion it should be elucidated here.