Fix indexing typos
There were some inconsistencies about the indexing of the examples. In
some places the txid starting with "91261eaf..." was indexed with a 1,
but it was reported as the first element of the array returned by the
command `bitcoin-cli listunspent` (an array with index starting from 0)
and later, in the same chapter, the same txid was reported as the
element with index 0 of the array returned by the same command.

Searching in text the txid I deduced that the correct indexing for that
txid was 1. So I made changes accordingly.