Fix broken link
The linked chapter (8.1) was neither about `redeemScript` nor was the link working (it lead to a 404). I checked and saw that chapter 10 is the one actually mentioning P2SH, and I assumed the best file was 10_3 because it is the one which more closely resembles "Building a Bitcoin Script with P2SH" (which is the linked yet broken file name) and also it mentions and deals with `redeemScript`.