Adding 'verbose=true' statement
I suggest to also add the `'verbose=true'` statement to the inspection of the legacy address in line 209. This way, even someone who does not remember what a legacy address in the testnet looks like, will get the additional information of `"type": "pubkeyhash"` so it's easier to see that it is indeed a legacy address and not segwit.